Move to ledger, change public key to composite key
